Restructure of directories
