--- index.ts
@@ -1,6 +1,0 @@
-import app from './src/config/config';
-
-
-app.listen(app.get("port"), () => {
-    console.log(`Servidor escuchando en el puerto ${app.get("port")}`);
-});
