+++ src/config/config.js
@@ -1,17 +1,18 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-const express_1 = __importDefault(require("express"));
-const path_1 = __importDefault(require("path"));
-const dotenv_1 = __importDefault(require("dotenv"));
-const routes_1 = __importDefault(require("../routes/routes"));
-dotenv_1.default.config();
-const app = (0, express_1.default)();
+const express = require('express');
+const path = require('path');
+const dotenv = require('dotenv');
+
+const pokedexRoutes = require('../routes/routes');
+
+dotenv.config();
+const app = express();
+
 const port = process.env.PORT || 3000;
 app.set('port', port);
+
 app.set('view engine', 'ejs');
-app.set('views', path_1.default.join(__dirname, '../resources/views'));
-app.use('/', routes_1.default);
-exports.default = app;
+app.set('views', path.join(__dirname, '../resources/views'));
+
+app.use('/', pokedexRoutes);
+
+module.exports = app;
