+++ src/controllers/api/pokedexApiController.js
@@ -1,13 +1,7 @@
-import { Request, Response } from 'express';
-import Pokedex from '../models/Pokedex';
+const Pokedex = require('../../models/Pokedex');
 
-class PokedexController {
-
-    public static index(req: Request, res: Response): void {
-        res.json({ message: 'Hello, world!' });
-    }
-
-    public static getRandomPokedex(req: Request, res: Response): void {
+class PokedexApiController {
+    static getRandomPokedex(req, res) {
         try {
             const pokedexes = Pokedex.getPokedexes();
             const randomPokedex = pokedexes[Math.floor(Math.random() * pokedexes.length)];
@@ -17,23 +11,21 @@ class PokedexController {
                 randomPokedex.height,
                 randomPokedex.ability,
                 randomPokedex.image,
-                randomPokedex.philosophicalPhrase,
+                randomPokedex.philosophicalPhrase
             );
 
             res.json({
                 id: pokedex.getId(),
                 name: pokedex.getName(),
                 height: pokedex.getHeight(),
-                ability: pokedex.getHeight(),
-                //containerId: process.env.HOSTNAME
+                ability: pokedex.getAbility(),
+                // containerId: process.env.HOSTNAME
             });
         } catch (error) {
             console.error(error);
             res.status(500).json({ error: 'Error while fetching the random Pokedex' });
         }
     }
-
-
 }
 
-export default PokedexController;
+module.exports = PokedexApiController;
