--- src/database/pokedexes.ts
@@ -1,84 +1,0 @@
-const pokedexData = [
-    {
-        "id": 1,
-        "name": "Bulbasaur",
-        "height": 0.7,
-        "ability": "Overgrow",
-        "image": "bulbasaur.png",
-        "philosophicalPhrase": "Life is like a garden, cultivate your growth."
-    },
-    {
-        "id": 2,
-        "name": "Ivysaur",
-        "height": 1.0,
-        "ability": "Overgrow",
-        "image": "ivysaur.png",
-        "philosophicalPhrase": "When the bud on its back starts swelling, a sweet aroma wafts to indicate the flower's coming bloom."
-    },
-    {
-        "id": 3,
-        "name": "Venusaur",
-        "height": 2.0,
-        "ability": "Overgrow",
-        "image": "venusaur.png",
-        "philosophicalPhrase": "After a rainy day, the flower on its back smells stronger. The scent attracts other Pokémon."
-    },
-    {
-        "id": 4,
-        "name": "Charmander",
-        "height": 0.6,
-        "ability": "Blaze",
-        "image": "charmander.png",
-        "philosophicalPhrase": "The flame that burns at the tip of its tail is an indication of its emotions."
-    },
-    {
-        "id": 5,
-        "name": "Charmeleon",
-        "height": 1.1,
-        "ability": "Blaze",
-        "image": "charmeleon.png",
-        "philosophicalPhrase": "Tough fights could excite this Pokémon. When excited, it may blow out bluish-white flames."
-    },
-    {
-        "id": 6,
-        "name": "Charizard",
-        "height": 1.7,
-        "ability": "Blaze",
-        "image": "charizard.png",
-        "philosophicalPhrase": "Its wings can carry this Pokémon close to an altitude of 4,600 feet. It blows out fire at very high temperatures."
-    },
-    {
-        "id": 7,
-        "name": "Squirtle",
-        "height": 0.5,
-        "ability": "Torrent",
-        "image": "squirtle.png",
-        "philosophicalPhrase": "Shoots water at prey while in the water."
-    },
-    {
-        "id": 8,
-        "name": "Wartortle",
-        "height": 1.0,
-        "ability": "Torrent",
-        "image": "wartortle.png",
-        "philosophicalPhrase": "It is recognized as a symbol of longevity. If its shell has algae on it, that Wartortle is very old."
-    },
-    {
-        "id": 9,
-        "name": "Blastoise",
-        "height": 1.6,
-        "ability": "Torrent",
-        "image": "blastoise.png",
-        "philosophicalPhrase": "A brutal Pokémon with pressurized water jets on its shell. They are used for high-speed tackles."
-    },
-    {
-        "id": 10,
-        "name": "Caterpie",
-        "height": 0.3,
-        "ability": "Shield Dust",
-        "image": "caterpie.png",
-        "philosophicalPhrase": "Its short feet are tipped with suction pads that enable it to tirelessly climb slopes and walls."
-    }
-];
-
-export default pokedexData;
