+++ src/models/Pokedex.js
@@ -1,10 +1,18 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-const pokedexes_1 = __importDefault(require("../database/pokedexes"));
+const pokedexData = require("../database/pokedexes");
+
 class Pokedex {
+    /*
+     * Represents a Pokemon entry in the Pokedex.
+     * 
+     * Attributes:
+     * - id: Unique identifier of the entry.
+     * - name: Name of the Pokemon.
+     * - height: Height of the Pokemon in meters.
+     * - ability: Special ability of the Pokemon.
+     * - image: Path to the image representing the Pokemon.
+     * - philosophicalPhrase: Philosophical phrase associated with the Pokemon.
+    */
+
     constructor(id, name, height, ability, image, philosophicalPhrase) {
         this.id = id;
         this.name = name;
@@ -13,41 +21,54 @@ class Pokedex {
         this.image = image;
         this.philosophicalPhrase = philosophicalPhrase;
     }
+
     getId() {
         return this.id;
     }
+
     getName() {
         return this.name;
     }
+
     setName(name) {
         this.name = name;
     }
+
     getHeight() {
         return this.height;
     }
+
     setHeight(height) {
         this.height = height;
     }
+
     getAbility() {
         return this.ability;
     }
+
     setAbility(ability) {
         this.ability = ability;
     }
+
     getImage() {
         return this.image;
     }
+
     setImage(image) {
         this.image = image;
     }
+
     getPhilosophicalPhrase() {
         return this.philosophicalPhrase;
     }
+
     setPhilosophicalPhrase(philosophicalPhrase) {
         this.philosophicalPhrase = philosophicalPhrase;
     }
+
     static getPokedexes() {
-        return pokedexes_1.default;
+        return pokedexData;
     }
 }
-exports.default = Pokedex;
+
+module.exports = Pokedex;
